+++ src/pages/api/v1/user/interview-prep/sheet.ts
@@ -2,8 +2,11 @@ import { apiStatusCodes } from '@/constant';
 import { NextApiRequest, NextApiResponse } from 'next';
 import { sendAPIResponse } from '@/utils';
 import { connectDB } from '@/middlewares';
-import { markQuestionCompletedByUser } from '@/database';
-import { MarkQuestionCompletedRequestProps } from '@/interfaces';
+import { markQuestionCompletedByUser, getAllQuestionsByUser } from '@/database';
+import {
+  MarkQuestionCompletedRequestProps,
+  GetAllQuestionsRequestProps,
+} from '@/interfaces';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   try {
@@ -14,6 +17,8 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     switch (method) {
       case 'PATCH':
         return handleMarkQuestionCompleted(req, res);
+      case 'GET':
+        return handleGetAllQuestions(req, res);
       default:
         return res.status(apiStatusCodes.BAD_REQUEST).json(
           sendAPIResponse({
@@ -69,6 +74,42 @@ const handleMarkQuestionCompleted = async (
       sendAPIResponse({
         status: false,
         message: 'Failed to update question status',
+        error,
+      })
+    );
+  }
+};
+
+const handleGetAllQuestions = async (
+  req: NextApiRequest,
+  res: NextApiResponse
+) => {
+  const { userId } = req.query as unknown as GetAllQuestionsRequestProps;
+
+  try {
+    const { data, error } = await getAllQuestionsByUser(userId);
+
+    if (error || !data) {
+      return res.status(apiStatusCodes.INTERNAL_SERVER_ERROR).json(
+        sendAPIResponse({
+          status: false,
+          message: 'Failed to retrieve questions',
+        })
+      );
+    }
+
+    return res.status(apiStatusCodes.OKAY).json(
+      sendAPIResponse({
+        status: true,
+        data,
+        message: 'Questions retrieved successfully',
+      })
+    );
+  } catch (error) {
+    return res.status(apiStatusCodes.INTERNAL_SERVER_ERROR).json(
+      sendAPIResponse({
+        status: false,
+        message: 'Failed to retrieve questions',
         error,
       })
     );
